% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoFor.R
\name{InfoFor}
\alias{InfoFor}
\title{InfoForList}
\usage{
InfoFor(list)
}
\arguments{
\item{list}{list to investigate}
}
\value{
summary information for each object in list
}
\description{
Lists size and contents information for each object in list
}
\examples{
InfoFor(ls())

18 Mar 2020 M. J. Brodzik brodzik@colorado.edu 
Copyright (C) 2019 Regents of the University of Colorado

}
