% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepModelDir.R
\name{PrepModelDir}
\alias{PrepModelDir}
\title{PrepModelDir}
\usage{
PrepModelDir(folder)
}
\arguments{
\item{folder}{string name of directory where new structure will be created}
}
\value{
no return
}
\description{
Creates required output directory structure for classifier and regression models.
If any of the directories don't already exist they will be created.
}
\details{
#' 18 Mar 2020 M. J. Brodzik brodzik@colorado.edu 
Copyright (C) 2019 Regents of the University of Colorado
}
