% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepPixDir.R
\name{PrepPixDir}
\alias{PrepPixDir}
\title{PrepPixDir}
\usage{
PrepPixDir(folder, tsize)
}
\arguments{
\item{folder}{string name of directory where new structure will be created}
}
\value{
pix directory for plots
}
\description{
Creates required output directory structure for plots.
If any of the directories don't already exist they will be created.
}
\details{
5 Mar 2024 L. w. Stephenson  
Copyright (C) 2019 Regents of the University of Colorado
}
