% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepSCADir.R
\name{PrepSCADir}
\alias{PrepSCADir}
\title{PrepSCADir}
\usage{
PrepSCADir(folder, modis_version, tsize)
}
\arguments{
\item{folder}{string name of directory where new structure will be created}
}
\value{
SCA directory
}
\description{
Creates required output directory structure for SCA compare days tables.
If any of the directories don't already exist they will be created.
}
\details{
27 Feb 2024 L. w. Stephenson  
Copyright (C) 2019 Regents of the University of Colorado
}
