% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyExtent.R
\name{StudyExtent}
\alias{StudyExtent}
\title{Descriptive characteristics of ESPFusion study extents}
\usage{
StudyExtent(extentName)
}
\value{
List with descriptive characteristics for the requested study extent
}
\description{
This class stores descriptive information for ESPFusion study extents
}
