% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Env.R
\name{Env}
\alias{Env}
\title{Directory locations for ESPFusion files}
\usage{
Env()
}
\value{
A list of directory locations used by the downscaling system
}
\description{
This class stores required directory locations for the ESPFusion
processing system.
}
