% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Env.R
\name{Env}
\alias{Env}
\title{Directory locations for ESPFusion files}
\usage{
Env()
}
\value{
A list of directory locations used by the downscaling system
}
\description{
This class stores required directory locations for the ESPFusion
processing system.
}
\details{
Thanks to Fong Chun Chan for help with building/using packages
https://tinyheero.github.io/jekyll/update/2015/07/26/making-your-first-R-package.html

following method examples from here:
https://www.cyclismo.org/tutorial/R/s3Classes.html

18 Mar 2020 M. J. Brodzik brodzik@colorado.edu 
Copyright (C) 2019 Regents of the University of Colorado
}
