% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAcompare.R
\name{SCAcompare}
\alias{SCAcompare}
\title{Compares fractional snow-covered area (fSCA) images}
\usage{
SCAcompare(myC, myT, thresh)
}
\arguments{
\item{myC}{first image}

\item{myT}{second image}

\item{thresh}{difference threshold (how is this used?)}
}
\value{
stats list of comparison statistics
}
\description{
This function compares 2 fSCA images.
}
\details{
#' 18 Mar 2020 M. J. Brodzik brodzik@colorado.edu 
Copyright (C) 2019 Regents of the University of Colorado
}
