% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyExtent.R
\name{StudyExtent}
\alias{StudyExtent}
\title{Descriptive characteristics of ESPFusion study extents}
\usage{
StudyExtent(extentName)
}
\value{
List with descriptive characteristics for the requested study extent
}
\description{
This class stores descriptive information for ESPFusion study extents
}
\details{
#' 18 Mar 2020 M. J. Brodzik brodzik@colorado.edu 
Copyright (C) 2019 Regents of the University of Colorado
}
